refactor: comment update, delete auth 적용

+++ src/main/java/kpl/fiml/comment/application/CommentService.java
@@ -24,7 +24,7 @@ public class CommentService {
 
     @Transactional
     public CommentCreateResponse create(Long userId, Long noticeId, CommentCreateRequest request) {
-        User user = getUserById(userId);
+        User user = getUserByUserId(userId);
         Notice notice = getNoticeById(noticeId);
         Comment createdComment = commentRepository.save(request.toEntity(user, notice));
 
@@ -43,16 +43,24 @@ public class CommentService {
     @Transactional
     public CommentUpdateResponse update(Long id, Long userId, CommentUpdateRequest request) {
         Comment findComment = getById(id);
-        validateUser(userId, findComment.getUser());
+        User user = getUserByUserId(userId);
+
+        if (!user.isSameUser(findComment.getUser())) {
+            throw new IllegalArgumentException("댓글 작성자만 댓글 수정이 가능합니다.");
+        }
         findComment.updateContent(request.getContent());
 
-        return CommentUpdateResponse.of(findComment.getId(), request.getUserId(), findComment.getContent(), findComment.getCreatedAt(), findComment.getUpdatedAt());
+        return CommentUpdateResponse.of(findComment.getId(), userId, findComment.getContent(), findComment.getCreatedAt(), findComment.getUpdatedAt());
     }
 
     @Transactional
     public CommentDeleteResponse deleteById(Long id, Long userId) {
         Comment findComment = getById(id);
-        validateUser(userId, findComment.getUser());
+        User user = getUserByUserId(userId);
+
+        if (!user.isSameUser(findComment.getUser())) {
+            throw new IllegalArgumentException("댓글 작성자만 댓글 삭제가 가능합니다.");
+        }
         findComment.delete();
 
         return CommentDeleteResponse.of(findComment.getId());
@@ -69,7 +77,7 @@ public class CommentService {
                 .orElseThrow(() -> new IllegalArgumentException("해당 댓글이 존재하지 않습니다."));
     }
 
-    private User getUserById(Long userId) {
+    private User getUserByUserId(Long userId) {
         return userRepository.findByIdAndDeletedAtIsNull(userId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 사용자가 존재하지 않습니다."));
     }
